refactor: optimize code for windows, clean up from needless code
