+++ test/manual/multipage-watch-partials/src/views/partials/footer.html
@@ -1,1 +1,1 @@
-<p>FOOTER: change me!</p>
+<p>FOOTER: change me!</p>12345
